requirements.txt
streamlit
google-generativeai
python-dotenv
pdf2image
altair<5
poppler-utils
